\name{nes2004}
\alias{nes2004}
\docType{data}
\title{Example data for SensMiceDA}
\description{A subset of data from the 2004 American National Election Study. }
\usage{data(nes2004)}
\format{
  A data frame with 542 observations on the following 5 variables.
  \describe{
    \item{leftright}{respondent left-right self-placement on a 1(left)-10(right) scale}
    \item{male}{a factor with levels \code{female} and \code{male}}
    \item{age}{respondent age in years}
	\item{educ}{a factor indicating respondent level of education with levels \code{No HS Diploma}, \code{HS Diploma}, \code{Some College}, \code{BA or BS}, \code{Graduate Degree}}
	\item{hhincome2}{Respondent household income, coded as the dollar amount in the middle of the income range specified}
}
}
\references{
The National Election Studies <www.electionstudies.org>. THE ANES 2004 TIME SERIES STUDY [dataset]. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer and distributor]. 
}
\keyword{datasets}
